many fixes on swipeable and conversation list
