+++ features/conversation-list/PinnedConversations/conversation-list-pinned-conversations.styles.ts
@@ -1,12 +1,15 @@
+import { useConversationListStyles } from "@/features/conversation-list/conversation-list.styles";
 import { useAppTheme } from "@/theme/useAppTheme";
 
 export function useConversationListPinnedConversationsStyles() {
   const { theme } = useAppTheme();
 
+  const { screenHorizontalPadding } = useConversationListStyles();
+
   return {
     avatarSize: theme.layout.grid.getColumnWidth({
       totalColumns: 3,
-      horizontalPadding: theme.layout.screenPadding.horizontal,
+      horizontalPadding: screenHorizontalPadding,
       gap: theme.spacing.sm,
     }),
   } as const;
