+++ features/conversation-list/components/conversation-list-item/V3DMListItem.tsx
@@ -1,9 +1,8 @@
 import { Avatar } from "@/components/Avatar";
-import { ISwipeableRenderActionsArgs, Swipeable } from "@/components/swipeable";
-import {
-  DeleteSwipeableAction,
-  ToggleUnreadSwipeableAction,
-} from "@/features/conversation-list/components/conversation-list-item/conversation-list-item-swipeable";
+import { ISwipeableRenderActionsArgs } from "@/components/swipeable";
+import { ConversationListItemSwipeable } from "@/features/conversation-list/components/conversation-list-item/conversation-list-item-swipeable/conversation-list-item-swipeable";
+import { ToggleUnreadSwipeableAction } from "./conversation-list-item-swipeable/conversation-list-item-swipeable-toggle-read-action";
+import { DeleteSwipeableAction } from "./conversation-list-item-swipeable/conversation-list-item-swipeable-delete-action";
 import { useHandleDeleteDm } from "@/features/conversation-list/hooks/useHandleDeleteDm";
 import { useConversationIsUnread } from "@/features/conversation-list/hooks/useMessageIsUnread";
 import { useMessageText } from "@/features/conversation-list/hooks/useMessageText";
@@ -86,13 +85,9 @@ export const V3DMListItem = ({ conversation }: V3DMListItemProps) => {
   });
 
   return (
-    <Swipeable
-      closeOnOpen
+    <ConversationListItemSwipeable
       renderRightActions={renderRightActions}
       renderLeftActions={renderLeftActions}
-      leftThreshold={theme.spacing["5xl"]}
-      leftActionsBackgroundColor={theme.colors.fill.caution}
-      rightActionsBackgroundColor={theme.colors.fill.minimal}
       onLeftSwipe={deleteDm}
       onRightSwipe={toggleReadStatus}
     >
@@ -104,6 +99,6 @@ export const V3DMListItem = ({ conversation }: V3DMListItemProps) => {
         subtitle={subtitle}
         isUnread={isUnread}
       />
-    </Swipeable>
+    </ConversationListItemSwipeable>
   );
 };
