+++ features/conversation-list/components/conversation-list-item/conversation-list-item-swipeable/conversation-list-item-swipeable-delete-action.tsx
@@ -1,0 +1,19 @@
+import { ISwipeableRenderActionsArgs } from "@/components/swipeable";
+import { useAppTheme } from "@/theme/useAppTheme";
+import React, { memo } from "react";
+import { ConversationListItemSwipeableAction } from "./conversation-list-item-swipeable-action";
+
+export const DeleteSwipeableAction = memo(
+  (props: ISwipeableRenderActionsArgs) => {
+    const { theme } = useAppTheme();
+
+    return (
+      <ConversationListItemSwipeableAction
+        icon="trash"
+        color={theme.colors.fill.caution}
+        actionProgress={props.progressAnimatedValue}
+        iconColor={theme.colors.fill.inverted.primary}
+      />
+    );
+  }
+);
