+++ features/conversation-list/components/conversation-list-item/conversation-list-item-swipeable/conversation-list-item-swipeable.styles.tsx
@@ -1,0 +1,8 @@
+import { useAppTheme } from "@/theme/useAppTheme";
+
+export function useConversationListItemSwipeableStyles() {
+  const { theme } = useAppTheme();
+  return {
+    swipeThreshold: theme.spacing["5xl"],
+  };
+}
