+++ features/conversation-list/conversation-list.styles.ts
@@ -1,0 +1,9 @@
+import { useAppTheme } from "@/theme/useAppTheme";
+
+export function useConversationListStyles() {
+  const { theme } = useAppTheme();
+
+  return {
+    screenHorizontalPadding: theme.spacing.lg,
+  };
+}
