+++ theme/layout.ts
@@ -1,5 +1,4 @@
 import { Dimensions } from "react-native";
-import { spacing } from "./spacing";
 
 const window = {
   width: Dimensions.get("window").width,
@@ -10,12 +9,6 @@ export const layout = {
   screen: {
     width: window.width,
     height: window.height,
-  },
-
-  // Common screen patterns
-  screenPadding: {
-    horizontal: spacing.lg,
-    onboarding: spacing.lg,
   },
 
   // Grid systems
